Stage 5 of 5 for JetBrains Academy - Kotlin - Connect Four project.
For ConnectFour.kt and GameBoard.kt I replaced named lambdas with functions ( I found out using lambdas as functions adds overhead, so I'm going to try and limit their use ).

ConnectFour.kt - changes:
• added Int totalGames to the constructor
• isMultiple - Boolean that informs if there will be more than one game
• score - variable of Int pairs to keep track of each player's score ( if it isMultiple is true ).
• isPlayer1 and command - moved from play()
• currentPlayer - used to get the current player using a getter that returns a player based on the value of isPlayer1
• init was adjusted to no longer print the gameBoard, and instead call printColumnRows() of gameBoard; and then print an appropriate message depending on if it's a single game or multiple game.
play() - with the exception of the last line, the contents were moved to playGame(). It now just makes a call to playGame() and then prints the game over message.
playGame() - it repeats the actions of what was in play() the number of times the players want to play the game. Lambda values turnMessage, getDisc, and getCommand were removed. Before the while loop the head for a multi game is printed with game number ( if isMultiple is true ); and then the gameBoard is printed. The call to getCommand within the while loop check was replaced with isContinue(). The call to insertDisc() of the gameBoard was adjusted to what was the contents of the getDisc lambda. The setting of drawOrWin for a win condition was adjusted to make use of the currentPlayer value. Once the while loop is complete it checks to confirm that the gameBoard gameOver() is true, if not it'll return to the calling function ( which would mean the user typed "end" ). A call to multipleGame() was added at the end ( which will only get called for a multi game ).
• isContinue() - adjusted version of getCommand, where instead of returning the command, it returns true or false depending on if the user typed "end"
• multipleGame() - sets the scores for each user dependent on if the current user won or if it was a draw. The score is then printed, and a call to resetGame() of gameBoard is made.
• companion object - the return is adjusted to return the value for totalGames using getGameTotal(). getGameTotal() gets the total games from the user, returning 1 if the user pressed enter, or the number for any number that is greater than zero. If the user types in an invalid number it prints an error and then returns itself to attempt to get the total games from the user again.